requirements.txt
folium
google-api-python-client
earthengine-api
django-geojson
django-leaflet